+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The types of changes are:
 - Page to edit or add privacy notices [#3058](https://github.com/ethyca/fides/pull/3058)
 - Side navigation bar can now also have children navigation links [#3099](https://github.com/ethyca/fides/pull/3099)
 - Endpoints for consent reporting [#3095](https://github.com/ethyca/fides/pull/3095)
+- Custom fields table [#3097](https://github.com/ethyca/fides/pull/3097)
 
 ### Changed
 - The `cursor` pagination strategy now also searches for data outside of the `data_path` when determining the cursor value [#3068](https://github.com/ethyca/fides/pull/3068)